refactor: split ports/grpc/ into cmd/, adapters/grpc_control/, adapters/webrtc/
ports/grpc/ mixed two unrelated transports (gRPC signaling client and
WebRTC peer connection + H.264 codec) with the binary entry point.
They are now separated by concern:

  cmd/accelerator_control_client/main.cpp
    ← binary entry point (was ports/grpc/main_client.cpp)

  adapters/grpc_control/
    ← outbound mTLS gRPC control client
    (accelerator_control_client, processor_engine_adapter, processor_engine_provider)

  adapters/webrtc/
    ← WebRTC media path: peer sessions, H.264 codec, data-channel framing
    (webrtc_manager, live_video_processor, data_channel_framing)

Namespace changes:
  jrb::ports::grpc_service → jrb::adapters::grpc_control (gRPC files)
  jrb::ports::grpc_service → jrb::adapters::webrtc (WebRTC files)

New BUILD files: cmd/accelerator_control_client/, adapters/grpc_control/,
adapters/webrtc/. Old ports/grpc/BUILD deleted.
