+++ src/cpp_accelerator/adapters/webrtc/webrtc_manager.h
@@ -15,14 +15,14 @@
 
 #include "proto/_virtual_imports/image_processor_service_proto/image_processor_service.pb.h"
 #include "src/cpp_accelerator/adapters/compute/cuda/memory/cuda_memory_pool.h"
-#include "src/cpp_accelerator/ports/grpc/data_channel_framing.h"
-#include "src/cpp_accelerator/ports/grpc/live_video_processor.h"
+#include "src/cpp_accelerator/adapters/webrtc/data_channel_framing.h"
+#include "src/cpp_accelerator/adapters/webrtc/live_video_processor.h"
 
 namespace jrb::application::engine {
 class ProcessorEngine;
 }
 
-namespace jrb::ports::grpc_service {
+namespace jrb::adapters::webrtc {
 
 class WebRTCManager : public std::enable_shared_from_this<WebRTCManager> {
 public:
@@ -95,4 +95,4 @@ public:
   std::thread cleanup_thread_;
 };
 
-}  // namespace jrb::ports::grpc_service
+}  // namespace jrb::adapters::webrtc
