+++ src/cpp_accelerator/cmd/accelerator_control_client/BUILD
@@ -1,0 +1,20 @@
+load("@rules_cc//cc:defs.bzl", "cc_binary")
+
+cc_binary(
+    name = "accelerator_control_client",
+    srcs = ["main.cpp"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/cpp_accelerator/adapters/grpc_control:accelerator_control_client",
+        "//src/cpp_accelerator/adapters/grpc_control:processor_engine_adapter",
+        "//src/cpp_accelerator/adapters/webrtc:webrtc_manager",
+        "//src/cpp_accelerator/application/engine:engine",
+        "//src/cpp_accelerator/core:signal_handler",
+        "//src/cpp_accelerator/core:version",
+        "//proto:image_processor_service_cc_proto",
+        "@com_github_grpc_grpc//:grpc++",
+        "@spdlog",
+        "@com_google_absl//absl/flags:flag",
+        "@com_google_absl//absl/flags:parse",
+    ],
+)
